+++ CHANGELOG.md
@@ -6,8 +6,10 @@ All notable changes to this project will be documented in this file.
 ### [Unreleased]
 
 #### Added
+- Rule to lint for [unused variables]
+(https://github.com/host-anshu/ansible-lint-rules/blob/master/rules/specs/unused_vars.feature).
 - Suppressing stderr output as well.
-- Tests for the tool utilities.
+- Tests for the tool utilities and runner partially.
 
 #### Fixed
 - Error for cases when worker wasn't instantiated due to any error.